da05完成用户端登陆和商品浏览接口

+++ sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -110,7 +110,7 @@ public class CategoryController {
      */
     @ApiOperation("根据类型获取分类信息接口")
     @GetMapping("/list")
-    public Result list(Integer type){
+    public Result list(Long type){
         List<Category> categoryList =categoryService.listType(type);
         return Result.success(categoryList);
     }
